Paper and question get corrected

+++ Modules/ScholarFlow.Modules.Academic/Commands/Papers/AddQuestionToPaper/AddQuestionToPaperCommand.cs
@@ -10,7 +10,7 @@ public sealed record AddQuestionToPaperCommand(
     string? QuestionImageUrl,
     int OrderIndex,
     DifficultyLevel? ManualDifficulty,
-    int Marks,
+    decimal Marks,
     List<AddOptionItem> Options) : IRequest<Guid>;
 
 public sealed record AddOptionItem(
