+++ Modules/ScholarFlow.Modules.Academic/Commands/Papers/UpdateQuestion/UpdateQuestionCommand.cs
@@ -9,5 +9,5 @@ public sealed record UpdateQuestionCommand(
     string QuestionText,
     string? QuestionImageUrl,
     int OrderIndex,
-    int Marks,
+    decimal Marks,
     DifficultyLevel? ManualDifficulty) : IRequest;
